apps/server/src/students/infrastructure/students.controller.ts
import { Controller, Get, Post, Body, Put, Param, Delete, Query, ParseIntPipe } from '@nestjs/common';
import { StudentsService } from './students.service';
import { CreateStudentDto } from './dto/create-student.dto';
import { UpdateStudentDto } from './dto/update-student.dto';
import { CriteriaDto } from '../../common/infrastructure/dto/criteria.dto';

@Controller('students')
export class StudentsController {
  constructor(private readonly studentsService: StudentsService) { }

  @Post()
  create(@Body() createTeacherDto: CreateStudentDto) {
    return this.studentsService.create(createTeacherDto);
  }

  @Get()
  findAll(
    @Query() criteriaDto: CriteriaDto
  ) {
    return this.studentsService.findAll(criteriaDto);
  }

  @Get(':id')
  findOne(@Param('id', ParseIntPipe) id: number) {
    return this.studentsService.findOne(id);
  }

  @Put(':id')
  update(@Param('id', ParseIntPipe) id: number, @Body() updateTeacherDto: UpdateStudentDto) {
    return this.studentsService.update(id, updateTeacherDto);
  }

  @Delete(':id')
  remove(@Param('id', ParseIntPipe) id: number) {
    return this.studentsService.remove(id);
  }
}
